refactor(frontend): remove AuthForm, modified hooks related to login/signup
