+++ packages/frontend/src/hooks/useLoginWithWallet.ts
@@ -1,35 +1,21 @@
-import { UserState } from '@unirep/core'
-import LOGIN_ERROR_MESSAGES from '../constants/error-messages/loginErrorMessage'
+import { useNavigate } from 'react-router-dom'
+import { useUser } from '../contexts/User'
+import { LocalStorageHelper } from '../utils/LocalStorageHelper'
 
-const useLoginWithWallet = (
-    accessToken: string | null,
-    hashUserId: string | null,
-    navigate: (path: string) => void,
-    setErrorCode: (errorCode: keyof typeof LOGIN_ERROR_MESSAGES) => void,
-    handleWalletSignMessage: (hashUserId: string) => Promise<void>,
-    setIsLogin: (param: string) => void,
-    createUserState: () => Promise<UserState>
-) => {
+export function useLoginWithWallet() {
+    const navigate = useNavigate()
+    const {
+        handleWalletSignMessage,
+        setIsLogin,
+        createUserState,
+        setErrorCode,
+    } = useUser()
     const loginWithWallet = async () => {
         try {
-            if (!hashUserId) {
-                throw new Error(LOGIN_ERROR_MESSAGES.MISSING_ELEMENT.code)
-            }
-            localStorage.setItem('hashUserId', hashUserId)
-            if (!accessToken) {
-                throw new Error(LOGIN_ERROR_MESSAGES.MISSING_ELEMENT.code)
-            }
-            localStorage.setItem('token', accessToken)
-            if (!window.ethereum) {
-                throw new Error(LOGIN_ERROR_MESSAGES.NO_WALLET.code)
-            }
-            try {
-                await handleWalletSignMessage(hashUserId)
-            } catch (error: any) {
-                throw new Error(LOGIN_ERROR_MESSAGES.WALLET_ISSUE.code)
-            }
+            const hashUserId = LocalStorageHelper.getGuaranteedHashUserId()
+            await handleWalletSignMessage(hashUserId)
             await createUserState()
-            localStorage.removeItem('showLogin')
+            LocalStorageHelper.removeIsTwitterVerified()
             setIsLogin('success')
             navigate('/')
         } catch (error: any) {
@@ -39,5 +25,3 @@ const useLoginWithWallet = (
 
     return loginWithWallet
 }
-
-export default useLoginWithWallet
