+++ packages/frontend/src/hooks/useSignupWithWallet.ts
@@ -1,55 +1,26 @@
-import { UserState } from '@unirep/core'
-import LOGIN_ERROR_MESSAGES from '../constants/error-messages/loginErrorMessage'
-import { SignupStatus } from '../contexts/User'
+import { useNavigate } from 'react-router-dom'
+import { useUser } from '../contexts/User'
+import { LocalStorageHelper } from '../utils/LocalStorageHelper'
 
-declare global {
-    interface Window {
-        ethereum: any
-    }
-}
-
-const useSignupWithWallet = (
-    accessToken: string | null,
-    hashUserId: string | null,
-    navigate: (path: string) => void,
-    setSignupStatus: (param: SignupStatus) => void,
-    setErrorCode: (errorCode: keyof typeof LOGIN_ERROR_MESSAGES) => void,
-    handleWalletSignMessage: (hashUserId: string) => Promise<void>,
-    signup: (
-        fromServer: boolean,
-        userStateInstance: UserState,
-        hashUserId: string,
-        accessToken: string
-    ) => Promise<void>,
-    setIsLogin: (param: string) => void,
-    createUserState: () => Promise<UserState>
-) => {
+export function useSignUpWithWallet() {
+    const navigate = useNavigate()
+    const {
+        handleWalletSignMessage,
+        signup,
+        setIsLogin,
+        createUserState,
+        setErrorCode,
+        setSignupStatus,
+    } = useUser()
     const signUpWithWallet = async () => {
         try {
-            if (!hashUserId) {
-                throw new Error(LOGIN_ERROR_MESSAGES.MISSING_ELEMENT.code)
-            }
-            localStorage.setItem('hashUserId', hashUserId)
-            if (!accessToken) {
-                throw new Error(LOGIN_ERROR_MESSAGES.MISSING_ELEMENT.code)
-            }
-            localStorage.setItem('token', accessToken)
-            if (!window.ethereum) {
-                throw new Error(LOGIN_ERROR_MESSAGES.NO_WALLET.code)
-            }
-            try {
-                await handleWalletSignMessage(hashUserId)
-            } catch (error: any) {
-                throw new Error(LOGIN_ERROR_MESSAGES.WALLET_ISSUE.code)
-            }
+            const accessToken = LocalStorageHelper.getGuaranteedAccessToken()
+            const hashUserId = LocalStorageHelper.getGuaranteedHashUserId()
+            await handleWalletSignMessage(hashUserId)
             const userStateInstance = await createUserState()
             setSignupStatus('pending')
             navigate('/')
-            try {
-                await signup(false, userStateInstance, hashUserId, accessToken)
-            } catch (error: any) {
-                throw new Error(LOGIN_ERROR_MESSAGES.SIGNUP_FAILED.code)
-            }
+            await signup(false, userStateInstance, hashUserId, accessToken)
             setSignupStatus('success')
             setIsLogin('success')
         } catch (error: any) {
@@ -57,8 +28,5 @@ const useSignupWithWallet = (
             setErrorCode(error.message)
         }
     }
-
     return signUpWithWallet
 }
-
-export default useSignupWithWallet
