+++ packages/frontend/src/pages/Signup/InternalSignup.tsx
@@ -14,8 +14,8 @@ import StepInfo from '../../components/login/StepInfo'
 import ErrorModal from '../../components/modal/ErrorModal'
 import NoteModal from '../../components/modal/NoteModal'
 import { useUser } from '../../contexts/User'
-import { useSignUpWithServer } from '../../hooks/useSignUpWithServer2'
-import { useSignUpWithWallet } from '../../hooks/useSignUpWithWallet2'
+import { useSignUpWithServer } from '../../hooks/useSignUpWithServer'
+import { useSignUpWithWallet } from '../../hooks/useSignUpWithWallet'
 import { LocalStorageHelper } from '../../utils/LocalStorageHelper'
 import { getVariantAutoScrollY } from '../../utils/motionVariants'
